Link in UPB by default for Bazel's python build rules.
In v21.x, we announced our intent to flip the default implementation of
protobuf python from pure python to upb.
- https://github.com/protocolbuffers/protobuf/releases/tag/v21.0

We also documented in docs and readme that
PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION env var can be set to override the
python implementation used.
- https://protobuf.dev/reference/python/python-generated/#sharing-messages

In practice, we only use upb if it can successfully be imported:
- https://github.com/protocolbuffers/protobuf/blob/1c29f34b24eafe4a362ce075c12c62c614fad199/python/google/protobuf/internal/api_implementation.py#L54
- https://github.com/protocolbuffers/protobuf/blob/1c29f34b24eafe4a362ce075c12c62c614fad199/python/google/protobuf/internal/api_implementation.py#L100

Otherwise, we fall back to pure python, even if the env var is set to upb.
Since we don't currently link in upb for Bazel, this means the default is
still pure python in practice and the environment variable is not respected.

This change links in UPB by default and ensures that default python builds and
unit tests using Bazel will enable running with python-UPB.

#test-continuous

PiperOrigin-RevId: 711436307